+++ src/main/api/cf/exposers.ts
@@ -2,6 +2,7 @@ import { ipcMain } from 'electron';
 
 import * as contests from './contests';
 import * as problems from './problems';
+import * as account from './account';
 
 ipcMain.handle('api:cf:getContestList', contests.getContestList);
 ipcMain.handle('api:cf:getHistoryContestList', contests.getHistoryContestInfo);
@@ -10,4 +11,9 @@ ipcMain.handle('api:cf:getContestProblem', async function (event, contestId) {
 });
 ipcMain.handle('api:cf:getProblemDetailedInfo', async function (event, prop) {
     return await problems.getProblemDetailedInfo(prop);
+});
+// Account
+ipcMain.handle('api:cf:logInCfAccount', async function (event, prop: account.logInCfAccountConfig) {
+    let accountInfo = await account.logInCfAccount(prop);
+    return accountInfo;
 });